pyproject.toml
# (C) Copyright 2024 Anemoi contributors.
#
# This software is licensed under the terms of the Apache Licence Version 2.0
# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
#
# In applying this licence, ECMWF does not waive the privileges and immunities
# granted to it by virtue of its status as an intergovernmental organisation
# nor does it submit to any jurisdiction.

[build-system]
requires = [
  "setuptools>=60",
  "setuptools-scm>=8",
]

[project]
name = "anemoi-datasets"

description = "A package to hold various functions to support training of ML models on ECMWF data."
keywords = [
  "ai",
  "datasets",
  "tools",
]

license = { file = "LICENSE" }
authors = [
  { name = "European Centre for Medium-Range Weather Forecasts (ECMWF)", email = "software.support@ecmwf.int" },
]

requires-python = ">=3.9"

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]

dynamic = [
  "version",
]
dependencies = [
  "anemoi-transform>=0.1.4",
  "anemoi-utils[provenance]>=0.4.16",
  "cfunits",
  "numpy",
  "pyyaml",
  "semantic-version",
  "tqdm",
  "zarr<=2.18.4",
]

optional-dependencies.all = [
  "anemoi-datasets[create,remote,xarray]",
]

optional-dependencies.create = [
  "earthkit-data[mars]>=0.12.4",
  "earthkit-geo>=0.3",
  "earthkit-meteo>=0.3",
  "eccodes>=2.39.1",
  "entrypoints",
  "pyproj>=3",
]

optional-dependencies.dev = [
  "anemoi-datasets[all,docs,tests]",
]

optional-dependencies.docs = [
  "nbsphinx",
  "pandoc",
  "sphinx",
  "sphinx-argparse",
  "sphinx-rtd-theme",
]

optional-dependencies.remote = [
  "boto3",
  "requests",
]

optional-dependencies.tests = [
  "anemoi-datasets[xarray]",
  "pytest",
]

optional-dependencies.xarray = [
  "adlfs",
  "gcsfs",
  "kerchunk",
  "pandas",
  "planetary-computer",
  "pystac-client",
]

urls.Documentation = "https://anemoi-datasets.readthedocs.io/"
urls.Homepage = "https://github.com/ecmwf/anemoi-datasets/"
urls.Issues = "https://github.com/ecmwf/anemoi-datasets/issues"
urls.Repository = "https://github.com/ecmwf/anemoi-datasets/"
# Changelog = "https://github.com/ecmwf/anemoi-datasets/CHANGELOG.md"

scripts.anemoi-datasets = "anemoi.datasets.__main__:main"

[tool.setuptools.package-data]
"anemoi.datasets.data" = [
  "*.css",
]

[tool.setuptools_scm]
version_file = "src/anemoi/datasets/_version.py"

[tool.isort]
profile = "black"

[tool.mypy]
strict = false
exclude = [
  "docs/**",
]

[tool.pydocstringformatter]
write = true
exclude = "docs/**"
style = "numpydoc"
